backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
motor>=3.1.1
pydantic[email]>=1.10.7
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0